auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt,argon2]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
redis==5.0.1
prometheus-client==0.19.0
python-json-logger==2.0.7
httpx==0.25.2
tenacity==8.2.3

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# OAuth 2.0
authlib==1.3.0
requests==2.31.0

# SAML 2.0
python3-saml==1.16.0
xmlsec==1.3.13
lxml==4.9.3
pycryptodome==3.19.0

# Multi-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2
twilio==8.10.0

# Security
cryptography==41.0.7
argon2-cffi==23.1.0
secure==0.3.0
user-agents==2.2.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Rate limiting
slowapi==0.1.9

# Session management
itsdangerous==2.1.2

# Device fingerprinting
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
freezegun==1.2.2
faker==20.1.0
factory-boy==3.3.0
httpx==0.25.2